// web/src/api/openapi-schema/driveCredentialsStatus.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * storyden
 * Storyden social API for building community driven platforms.
The Storyden API does not adhere to semantic versioning but instead applies a rolling strategy with deprecations and minimal breaking changes. This has been done mainly for a simpler development process and it may be changed to a more fixed versioning strategy in the future. Ultimately, the primary way Storyden tracks versions is dates, there are no set release tags currently.

 * OpenAPI spec version: v1.26.13-post
 */
import type { DriveCredentialsSource } from "./driveCredentialsSource";

export interface DriveCredentialsStatus {
  configured: boolean;
  /** The service account's email address. Share a Google Drive folder
with this address to make it reachable. Absent when not
configured.
 */
  service_account_email?: string;
  source: DriveCredentialsSource;
}